+++ nilearn/decoding/tests/test_searchlight.py
@@ -4,23 +4,21 @@ import numpy as np
 import pytest
 from nibabel import Nifti1Image
 from sklearn.model_selection import KFold, LeaveOneGroupOut
-from sklearn.utils.estimator_checks import parametrize_with_checks
 
 from nilearn._utils.estimator_checks import (
     check_estimator,
-    return_expected_failed_checks,
 )
 from nilearn._utils.tags import SKLEARN_LT_1_6
 from nilearn.conftest import _img_3d_ones, _rng
 from nilearn.decoding import searchlight
 
+ESTIMATOR_TO_CHECK = [searchlight.SearchLight(mask_img=_img_3d_ones())]
+
 if SKLEARN_LT_1_6:
 
     @pytest.mark.parametrize(
         "estimator, check, name",
-        check_estimator(
-            estimators=[searchlight.SearchLight(mask_img=_img_3d_ones())],
-        ),
+        check_estimator(estimators=ESTIMATOR_TO_CHECK),
     )
     def test_check_estimator_sklearn_valid(estimator, check, name):  # noqa: ARG001
         """Check compliance with sklearn estimators."""
@@ -29,24 +27,21 @@ if SKLEARN_LT_1_6:
     @pytest.mark.xfail(reason="invalid checks should fail")
     @pytest.mark.parametrize(
         "estimator, check, name",
-        check_estimator(
-            estimators=[searchlight.SearchLight(mask_img=_img_3d_ones())],
-            valid=False,
-        ),
+        check_estimator(estimators=ESTIMATOR_TO_CHECK, valid=False),
     )
     def test_check_estimator_sklearn_invalid(estimator, check, name):  # noqa: ARG001
         """Check compliance with sklearn estimators."""
         check(estimator)
 
-else:
+# else:
 
-    @parametrize_with_checks(
-        estimators=[searchlight.SearchLight(mask_img=_img_3d_ones())],
-        expected_failed_checks=return_expected_failed_checks,
-    )
-    def test_check_estimator_sklearn(estimator, check):
-        """Check compliance with sklearn estimators."""
-        check(estimator)
+#     @parametrize_with_checks(
+#         estimators=ESTIMATOR_TO_CHECK,
+#         expected_failed_checks=return_expected_failed_checks,
+#     )
+#     def test_check_estimator_sklearn(estimator, check):
+#         """Check compliance with sklearn estimators."""
+#         check(estimator)
 
 
 def _make_searchlight_test_data(frames):
