+++ nilearn/regions/tests/test_region_extractor.py
@@ -97,6 +97,22 @@ if SKLEARN_LT_1_6:
         """Check compliance with sklearn estimators."""
         check(estimator)
 
+# else:
+
+#     from sklearn.utils.estimator_checks import parametrize_with_checks
+
+#     from nilearn._utils.estimator_checks import (
+#         return_expected_failed_checks,
+#     )
+
+#     @parametrize_with_checks(
+#         estimators=ESTIMATORS_TO_CHECK,
+#         expected_failed_checks=return_expected_failed_checks,
+#     )
+#     def test_check_estimator_sklearn(estimator, check):
+#         """Check compliance with sklearn estimators."""
+#         check(estimator)
+
 
 @pytest.mark.parametrize(
     "estimator, check, name",
